% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_leaf_piles.R
\name{create_leaf_pile}
\alias{create_leaf_pile}
\title{Create a pile of leaves}
\usage{
create_leaf_pile(param)
}
\arguments{
\item{param}{A list of parameters specifying how to grow the leaf and how many leaves to grow}
}
\value{
A data frame containing an x, y, and leaf_id column
}
\description{
Create a pile of leaves
}
\examples{
get_ginkgo_params() \%>\% create_leaf_pile()

}
