% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_good_parameter_vals.R
\name{get_params}
\alias{get_params}
\title{Create a parameter list}
\usage{
get_params(
  n_leaves = 1,
  n_layer = 3,
  scale = c(0.8, 0.9),
  angle = c(-30, 10, 20),
  split = 3
)
}
\arguments{
\item{n_leaves}{The number of leaves to generate}

\item{n_layer}{How many layers to grow}

\item{scale}{A single value or vector of scaling length options}

\item{angle}{A single value or vector of angles to split from}

\item{split}{A single value or vector of the number of splits at each branch}
}
\value{
A list of parameter values
}
\description{
Create a parameter list
}
