% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_grow_leaves.R
\name{grow_leaf_layers}
\alias{grow_leaf_layers}
\title{Grow one layer of the leaf}
\usage{
grow_leaf_layers(leaf, param, n_layer)
}
\arguments{
\item{leaf}{The leaf in process of being grown}

\item{param}{The parameters specifying how to grow the leaf}

\item{n_layer}{Keeps track of the growth}
}
\value{
A tibble containing the growth at that layer, one row per branch
}
\description{
Leafs grow outward from the stem in layers. At each layer, we branch at each endpoint
}
\details{
Calls \code{one_branch()} at each of the endpoints in the current layer
}
