% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_types.R
\name{mid_branch}
\alias{mid_branch}
\title{Split in the middle of a branch}
\usage{
mid_branch(partial_leaf, param, n_layer)
}
\arguments{
\item{partial_leaf}{The leaf in progress of being grown}

\item{param}{The parameters specifying how to grow the leaf}

\item{n_layer}{Keeps track of the growth}
}
\value{
A one row tibble that represents the new branch


}
\description{
Counterpoint to \code{one_branch} which splits in the middle of the branch instead of the end
}
