% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_grow_leaves.R
\name{one_branch}
\alias{one_branch}
\title{Split the end of a branch}
\usage{
one_branch(partial_leaf, param, n_layer)
}
\arguments{
\item{partial_leaf}{The leaf in progress of being grown}

\item{param}{The parameters specifying how to grow the leaf}

\item{n_layer}{Keeps track of the growth}
}
\value{
The leaf with the new growth of splits
}
\description{
At each endpoint a leaf is growing from, split/branch depending on the parameters
This is the one iteration that is repeated on each endpoint of each leaf layer
}
