% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04a_move_leaves.R
\name{place_leaf}
\alias{place_leaf}
\title{Randomly place a leaf}
\usage{
place_leaf(leaf)
}
\arguments{
\item{leaf}{A single leaf grown from the origin}
}
\value{
A single leaf randomly placed
}
\description{
Since the timing of this method is slower than randomly placing leaves, it is discouraged to use this function
}
