% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot_leaves.R
\name{plot_leaves}
\alias{plot_leaves}
\title{Plot leaves}
\usage{
plot_leaves(
  leaves,
  pal = c("#997C67", "#755330", "#B0703C", "#DBA72E", "#E3CCA1", "#D1CEC5")
)
}
\arguments{
\item{leaves}{A list containing}

\item{pal}{A color palette for the leaves}
}
\value{
A ggplot object
}
\description{
Plots leaves using ggplot2, with \code{geom_path()} connecting each of the growth steps in order to grow leaf/tree objects
}
