% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{radians}
\alias{radians}
\title{Helper function to convert degrees into radians}
\usage{
radians(degree)
}
\arguments{
\item{degree}{}
}
\value{
The radian equivalent of the degree parameter
}
\description{
R's cos() and sin() functions take radian parameter values,
but personally its easier for me to think in degrees
}
