% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_rake_leaves.R
\name{rake_leaves}
\alias{rake_leaves}
\title{Rake leaves - Convert leaf to tidy data format}
\usage{
rake_leaves(leaf)
}
\arguments{
\item{leaf}{The data frame that stores the generated leaf}
}
\value{
A data frame representation that is easier to be passed to ggplot to plot, with an x and y column
}
\description{
The data format needed to grow the leaf using purrr's accumulate is not conducive to plotting the leaves, this function cleans up the data so it is more easily plotted
}
