% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04a_move_leaves.R
\name{rotate_leaf}
\alias{rotate_leaf}
\title{Rotate leaves grown from the origin}
\usage{
rotate_leaf(leaf)
}
\arguments{
\item{leaf}{A single leaf grown from the origin}
}
\value{
A single leaf that is randomly rotated around the origin
}
\description{
Since the timing of this method is slower than randomly placing leaves, it is discouraged to use this function
}
