% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_leaf_piles.R
\name{spread_leaves}
\alias{spread_leaves}
\title{A different (slower) method for creating a pile of leaves}
\usage{
spread_leaves(param, ndistinct = 3)
}
\arguments{
\item{param}{A list of parameters specifying how to grow the leaf and how many leaves to grow}

\item{ndistinct}{The number of distinct trees to create}
}
\value{
A data frame containing an x, y, and leaf_id column
}
\description{
This method was intended to spead up the process of generating trees. This method creates a smaller number of distinct trees (since they look similar in the end), and then randomly replicates those trees, and then randomly rotates and places them. It ends up being slower in the end
}
