% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_grow_leaves.R
\name{grow_leaf}
\alias{grow_leaf}
\title{Grow ONE leaf}
\usage{
grow_leaf(param, init_location = "random", func = one_branch)
}
\arguments{
\item{param}{A list of parameter values that give the leaves their
shape}

\item{init_location}{A character of either "random" or "origin"}

\item{func}{The function to either branch at midpoint or endpoint, defaults endpoint}
}
\value{

}
\description{
This function is the highest level to create a single leaf.
}
\details{
It starts with an initial leaf, which is essentially just a stalk. Depending on the parameter values, this is either grown at the origin or a randomly placed location

It calls growth_leaf_layers and one_branch to recursively grow the tree
}
