% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot_leaves.R
\name{plot_leaves}
\alias{plot_leaves}
\title{Plot leaves}
\usage{
plot_leaves(
  leaves,
  connect = FALSE,
  pal = c("#CD8E27", "#C65729", "#993327", "#592821", "#DBA72E", "#D28F33")
)
}
\arguments{
\item{leaves}{A list containing}

\item{pal}{A color palette for the leaves}
}
\value{
A ggplot object
}
\description{
Plots leaves using ggplot2, with \code{geom_path()} connecting each of the growth steps in order to grow leaf/tree objects
}
\details{
Another good palette is c("#686724","#CD8E27","#C65729","#993327","#592821","#DBA72E")

c( "#997C67", "#755330","#B0703C", "#DBA72E", "#E3CCA1","#D1CEC5")
}
